refactor: utils and server
