+++ lib/Server.js
@@ -711,6 +711,21 @@ class Server {
     });
   }
 
+  showStatus() {
+    const suffix =
+      this.options.inline !== false || this.options.lazy === true
+        ? '/'
+        : '/webpack-dev-server/';
+    const uri = `${createDomain(this.options, this.listeningApp)}${suffix}`;
+
+    status(
+      uri,
+      this.options,
+      this.log,
+      this.options.stats && this.options.stats.colors
+    );
+  }
+
   listen(port, hostname, fn) {
     this.hostname = hostname;
 
@@ -876,21 +891,6 @@ class Server {
 
     // disallow
     return false;
-  }
-
-  showStatus() {
-    const suffix =
-      this.options.inline !== false || this.options.lazy === true
-        ? '/'
-        : '/webpack-dev-server/';
-    const uri = `${createDomain(this.options, this.listeningApp)}${suffix}`;
-
-    status(
-      uri,
-      this.options,
-      this.log,
-      this.options.stats && this.options.stats.colors
-    );
   }
 
   // eslint-disable-next-line
